prov/efa: Move EFA long opts to efa_shared.h
Allows multiple tests to use them
